Fully implemented abstract factory design pattern for basic matrices (CLEED_BASIC_MATRIX_* macros) and vectors (CLEED_VECTOR_* macros) complete with typedefs for a unified interface for the rfac and search when using either native, gsl or blas backends.
Includes some other general refactoring and a continued move away from
Numerical Recipe methods.

See: cleed_vector.h, cleed_real.h (previously real.h) and cleed_matrix.h
header files for macros, typedefs and inline functions relating to
vectors and matrices.

+++ src/rfac/rfrp.c
@@ -52,7 +52,7 @@ real rfac_rp(const real *eng, const real *e_int, const real *t_int, real vi)
   the_y_sum = 0.;
   rfac_sum = 0.;
   for(i_eng = 1;
-     ! IS_EQUAL_REAL(eng[i_eng], F_END_OF_LIST); i_eng ++)
+     ! IS_EQUAL_REAL(eng[i_eng], (real)F_END_OF_LIST); i_eng ++)
   {
     e_step = eng[i_eng] - eng[i_eng-1];
 
@@ -61,17 +61,17 @@ real rfac_rp(const real *eng, const real *e_int, const real *t_int, real vi)
         ( e_step * 0.5 * (t_int[i_eng] + t_int[i_eng-1]) ));
 
     Y_the = (real)(L / ( 1. + L*L*vi*vi));
-    the_y_sum += SQUARE(Y_the)*e_step;
+    the_y_sum += SQUARE((double)Y_the)*e_step;
 
     /* experimental Y function */
     L = (real)((e_int[i_eng] - e_int[i_eng-1]) /
         ( e_step * 0.5 * (e_int[i_eng] + e_int[i_eng-1]) ));
 
     Y_exp = (real)(L / ( 1. + L*L*vi*vi));
-    exp_y_sum += SQUARE(Y_exp)*e_step;
+    exp_y_sum += SQUARE((double)Y_exp)*e_step;
 
     /* compute R-factor sum */
-    rfac_sum += SQUARE(Y_the - Y_exp)*e_step;
+    rfac_sum += SQUARE((double)Y_the - (double)Y_exp)*e_step;
   }
 
   #ifdef CONTROL
